+++ Questao2/Model/FootballMatch.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Questao2.Model
+{
+    public class FootballMatch
+    {
+        public string team1 { get; set; }
+        public string team2 { get; set; }
+        public string team1goals { get; set; }
+        public string team2goals { get; set; }
+        public int year { get; set; }
+    }
+}
